+++ src/lib/common/StdMap.h
@@ -11,4 +11,4 @@
 
 #include <map>
 
-#include "stdpost.h"
+#include "StdPost.h"
