+++ src/lib/common/StdOStream.h
@@ -15,4 +15,4 @@
 #include <iostream>
 #endif
 
-#include "stdpost.h"
+#include "StdPost.h"
